Fix various memory leaks in replay bots
The records aren't cached in this code but the record file is parsed everytime again some info is needed.

Whenever LoadRecordFromFile was called all the record frames were stored in a new adt array, but often never used - and never closed. This caused the same record to be leaked in memory A LOT.

Next to that the recording of TeleportEntity calls using DHooks was overwriting and adding handles that were never closed.